+++ NEWS.md
@@ -8,10 +8,12 @@
  
 ## Test Updates
 
+ * Add tests for `cancel()`.
+
  * Under development:
  
    - Assert that `plan()` accepts argument `interrupts`.
- 
+
 
 # Version 0.9.0 [2025-04-14]